add game data
