+++ crates/crabe_framework/src/data/world/game_state.rs
@@ -1,25 +1,35 @@
-use crate::data::world::{Team, TeamColor};
-use serde::Serialize;
+use serde::{Serialize};
+use crate::data::world::TeamColor;
 
-/// The `GameState` struct represents the state of the SSL game, including the teams and which team is on the positive half of the field.
-#[derive(Serialize, Clone, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct GameState {
-    /// The `Team` struct representing our ally team.
-    pub ally: Team,
-    /// The `Team` struct representing the enemy team.
-    pub enemy: Team,
-    /// The color of the team that is on the positive half of the field.
-    pub positive_half: TeamColor,
+#[derive(Serialize, Copy, Clone, Debug, Eq, PartialEq)]
+#[serde(rename_all="camelCase")]
+pub enum GameState {
+    Halted(HaltedState),
+    Stopped(StoppedState),
+    Running(RunningState)
 }
 
-impl GameState {
-    /// Creates a new `GameState` with the given `team_color` as the team color for the ally team, and the opposite team color for the enemy team.
-    pub fn new(team_color: TeamColor) -> Self {
-        Self {
-            ally: Team::with_color(team_color),
-            enemy: Team::with_color(team_color.opposite()),
-            positive_half: team_color.opposite(),
-        }
-    }
+#[derive(Serialize, Copy, Clone, Debug, Eq, PartialEq)]
+#[serde(rename_all="camelCase")]
+pub enum HaltedState {
+    Halt,
+    Timeout
+}
+
+#[derive(Serialize, Copy, Clone, Debug, Eq, PartialEq)]
+#[serde(rename_all="camelCase")]
+pub enum StoppedState {
+    Stop,
+    PrepareKickoff,
+    PreparePenalty,
+    BallPlacement
+}
+
+#[derive(Serialize, Copy, Clone, Debug, Eq, PartialEq)]
+#[serde(rename_all="camelCase")]
+pub enum RunningState {
+    KickOff(TeamColor),
+    Penalty,
+    FreeKick,
+    Run
 }
